requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4
openai==1.44.0
requests==2.31.0
python-dotenv==1.0.1
click==8.1.7
beautifulsoup4==4.12.3
lxml==5.1.0
boto3==1.34.34
pandas==2.2.0
tqdm==4.66.1
httpx==0.27.2
minio==7.2.3
aiohttp==3.9.1
aiofiles==23.2.1
tenacity==8.2.3